git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@3787 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/os/typedefs.h
@@ -83,16 +83,17 @@ typedef unsigned long DWORD;
 #endif  /* __WX__ */
 
 #else /* !defined(_WIN32) */
+#include <stdint.h>
 #include <sys/types.h>
 typedef int8_t  int8;
 typedef int16_t int16;
 typedef int32_t int32;
 typedef int64_t int64;
 
-typedef u_int8_t  uint8;
-typedef u_int16_t uint16;
-typedef u_int32_t uint32;
-typedef u_int64_t uint64;
+typedef uint8_t  uint8;
+typedef uint16_t uint16;
+typedef uint32_t uint32;
+typedef uint64_t uint64;
 
 typedef int errno_t;
 
@@ -118,9 +119,9 @@ typedef TCHAR *LPTSTR;
 typedef const TCHAR *LPCTSTR;
 typedef bool BOOL;
 typedef unsigned char BYTE;
-typedef unsigned short WORD;
-typedef unsigned long DWORD;
-typedef long LONG;
+typedef uint16_t WORD;
+typedef uint32_t DWORD;
+typedef int32_t LONG;
 #if defined(PWS_LITTLE_ENDIAN)
 #define LOBYTE(w) ((BYTE)(w))
 #define HIBYTE(w) ((BYTE)(((WORD)(w) >> 8) & 0xFF))
@@ -138,7 +139,7 @@ typedef long LONG;
 #else
 #error "One of PWS_LITTLE_ENDIAN or PWS_BIG_ENDIAN must be defined before including typedefs.h"
 #endif
-typedef long LPARAM;
+typedef int32_t LPARAM;
 typedef unsigned int UINT;
 typedef int HANDLE;
 #define INVALID_HANDLE_VALUE HANDLE(-1)
